fix: better handling of windows paths
